Added the maven-jar-plugin to add the correct main class into the manifest file of the .jar file.
Added the maven-shade-plugin to bundle the dependencies into the main
.jar file.

Updated the converter to use the URL.openStream() to read the example
xlsx file as it was failing to load when packaged within a jar when
using URL.getPath().

Added log4j as a dependency and included a call to the
BasicConfigurator() within the main method so that logging will go to
standard out when the program is run as a .jar file.